texto_a_voz/requirements.txt
gradio
python-dotenv
elevenlabs
translate